Rename contact directory
